+++ src/Notification/Payment.php
@@ -2,18 +2,52 @@
 
 namespace Wearesho\Bobra\IPay\Notification;
 
+use Wearesho\Bobra\IPay;
 use Wearesho\Bobra\Payments;
 
 /**
  * Class Payment
  * @package Wearesho\Bobra\IPay\Notification
  */
-class Payment
+class Payment implements IPay\PaymentInterface
 {
-    use Payments\PaymentTrait;
+    use IPay\PaymentTrait;
 
-    public function __construct()
-    {
+    /** @var string */
+    protected $ident;
+
+    /** @var int */
+    protected $amount;
+
+    /** @var string */
+    protected $currency;
+
+    /** @var \DateTime */
+    protected $timestamp;
+
+    /** @var Payments\TransactionCollection */
+    protected $transactions;
+
+    public function __construct(
+        int $id,
+        string $ident,
+        int $status,
+        int $amount,
+        string $currency,
+        \DateTime $timestamp,
+        Payments\TransactionCollection $transactions,
+        string $salt,
+        string $sign
+    ) {
+        $this->id = $id;
+        $this->ident = $ident;
+        $this->status = $status;
+        $this->amount = $amount;
+        $this->currency = $currency;
+        $this->timestamp = $timestamp;
+        $this->transactions = $transactions;
+        $this->salt = $salt;
+        $this->sign = $sign;
     }
 
     /**
@@ -22,7 +56,49 @@ class Payment
      */
     public function getIdent(): string
     {
+        return $this->ident;
     }
 
+    /**
+     * @return Payments\TransactionCollection
+     */
+    public function getTransactions(): Payments\TransactionCollection
+    {
+        return $this->transactions;
+    }
 
+    /**
+     * Дата проведения последней операции по платежу
+     * @return \DateTime
+     */
+    public function getTimestamp(): \DateTime
+    {
+        return $this->timestamp;
+    }
+
+    /**
+     * 3х буквенных код валюты платежа
+     * @return string
+     */
+    public function getCurrency(): string
+    {
+        return $this->currency;
+    }
+
+    /**
+     * Общая сумма платежа, сумма всех транзакций и комиссии в копейках
+     * @return int
+     */
+    public function getAmount(): int
+    {
+        return $this->amount;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function jsonSerialize()
+    {
+        return $this;
+    }
 }